Move build_url to utils.py
